Moves pickers files + checkstyle
