+++ main/core/Resources/modules/data/types/users/components/display.jsx
@@ -8,27 +8,22 @@ import {EmptyPlaceholder} from '#/main/core/layout/components/placeholder'
 import {User as UserType} from '#/main/core/user/prop-types'
 import {UserCard} from '#/main/core/user/data/components/user-card'
 
-const UsersDisplay = (props) =>
+const UsersDisplay = (props) => !isEmpty(props.data) ?
   <div>
-    {!isEmpty(props.data) ?
-      <div>
-        {props.data.map(user =>
-          <UserCard
-            key={`user-card-${user.id}`}
-            data={user}
-            size="sm"
-            orientation="col"
-          />
-        )}
-      </div> :
-      <EmptyPlaceholder
-        size="lg"
-        icon="fa fa-user"
-        title={trans('no_user')}
+    {props.data.map(user =>
+      <UserCard
+        key={`user-card-${user.id}`}
+        data={user}
+        size="sm"
+        orientation="col"
       />
-    }
-  </div>
-
+    )}
+  </div> :
+  <EmptyPlaceholder
+    size="lg"
+    icon="fa fa-user"
+    title={trans('no_user')}
+  />
 
 UsersDisplay.propTypes = {
   data: T.arrayOf(T.shape(UserType.propTypes))
