+++ main/core/Resources/modules/modals/groups/components/modal.jsx
@@ -7,7 +7,7 @@ import {Modal} from '#/main/app/overlay/modal/components/modal'
 import {ListData} from '#/main/app/content/list/containers/data'
 
 import {trans} from '#/main/core/translation'
-import {selectors} from '#/main/core/modals/pickers/group/store'
+import {selectors} from '#/main/core/modals/groups/store'
 import {GroupList} from '#/main/core/administration/user/group/components/group-list'
 import {Group as GroupType} from '#/main/core/user/prop-types'
