+++ main/core/Resources/modules/modals/groups/store/index.js
@@ -1,0 +1,12 @@
+/**
+ * Groups picker modal store.
+ */
+
+import {reducer} from '#/main/core/modals/groups/store/reducer'
+import {selectors} from '#/main/core/modals/groups/store/selectors'
+
+// export store module
+export {
+  reducer,
+  selectors
+}
