+++ main/core/Resources/modules/modals/groups/store/selectors.js
