--- main/core/Resources/modules/modals/pickers/group/store/index.js
@@ -1,12 +1,0 @@
-/**
- * Groups picker modal store.
- */
-
-import {reducer} from '#/main/core/modals/pickers/group/store/reducer'
-import {selectors} from '#/main/core/modals/pickers/group/store/selectors'
-
-// export store module
-export {
-  reducer,
-  selectors
-}
