--- main/core/Resources/modules/modals/pickers/user/store/index.js
@@ -1,12 +1,0 @@
-/**
- * Users picker modal store.
- */
-
-import {reducer} from '#/main/core/modals/pickers/user/store/reducer'
-import {selectors} from '#/main/core/modals/pickers/user/store/selectors'
-
-// export store module
-export {
-  reducer,
-  selectors
-}
