+++ main/core/Resources/modules/modals/users/components/modal.jsx
@@ -7,7 +7,7 @@ import {Modal} from '#/main/app/overlay/modal/components/modal'
 import {ListData} from '#/main/app/content/list/containers/data'
 
 import {trans} from '#/main/core/translation'
-import {selectors} from '#/main/core/modals/pickers/user/store'
+import {selectors} from '#/main/core/modals/users/store'
 import {UserList} from '#/main/core/administration/user/user/components/user-list'
 import {User as UserType} from '#/main/core/user/prop-types'
