+++ main/core/Resources/modules/modals/users/store/index.js
@@ -1,0 +1,12 @@
+/**
+ * Users picker modal store.
+ */
+
+import {reducer} from '#/main/core/modals/users/store/reducer'
+import {selectors} from '#/main/core/modals/users/store/selectors'
+
+// export store module
+export {
+  reducer,
+  selectors
+}
